updated plugins for 2.89.1.7, replaced slack with hipchat
