+++ dockerBuildPublish.properties
@@ -1,3 +1,3 @@
-tag=2.89.1.7
+tag=test-1-2.89.1.7
 #allow this branch to get pushed to Docker Hub
-pushBranch=false
+pushBranch=true
