+++ plugins.txt
@@ -1,131 +1,133 @@
+ace-editor:1.1
+ansicolor:0.5.2
+antisamy-markup-formatter:1.5
+apache-httpcomponents-client-4-api:4.5.3-2.0
 async-http-client:1.7.24.1
-blueocean-commons:1.3.0
+authentication-tokens:1.3
+aws-credentials:1.21
+aws-java-sdk:1.11.119
+blueocean-autofavorite:1.1.0
+blueocean-bitbucket-pipeline:1.3.3
+blueocean-commons:1.3.3
+blueocean-config:1.3.3
+blueocean-dashboard:1.3.3
+blueocean-display-url:2.1.1
+blueocean-events:1.3.3
+blueocean-git-pipeline:1.3.3
+blueocean-github-pipeline:1.3.3
+blueocean-i18n:1.3.3
+blueocean-jira:1.3.3
+blueocean-jwt:1.3.3
+blueocean-personalization:1.3.3
+blueocean-pipeline-api-impl:1.3.3
+blueocean-pipeline-editor:1.3.3
+blueocean-pipeline-scm-api:1.3.3
+blueocean-rest-impl:1.3.3
+blueocean-rest:1.3.3
+blueocean-web:1.3.3
+blueocean:1.3.3
 bouncycastle-api:2.16.2
-cloudbees-license:9.17
-nectar-license:8.6
-structs:1.10
-credentials:2.1.16
-jackson2-api:2.7.3
-metrics:3.1.2.10
-support-core:2.41
-cloudbees-assurance:2.73.0.2
+branch-api:2.0.11
+cloudbees-assurance:2.89.0.1
+cloudbees-bitbucket-branch-source:2.2.3
 cloudbees-blueocean-default-theme:0.3
 cloudbees-folder:6.1.2
-cloudbees-folders-plus:3.2
-display-url-api:2.0
-durable-task:1.14
+cloudbees-folders-plus:3.3
+cloudbees-groovy-view:1.7
+cloudbees-jsync-archiver:5.6
+cloudbees-license:9.17
+cloudbees-monitoring:2.7
+cloudbees-nodes-plus:1.16
+cloudbees-ssh-slaves:2.1
+cloudbees-support:3.12
+cloudbees-template:4.34
+cloudbees-view-creation-filter:1.5
+cloudbees-workflow-template:2.7
+cloudbees-workflow-ui:2.1
+command-launcher:1.1
+credentials-binding:1.13
+credentials:2.1.16
+display-url-api:2.1.0
+docker-commons:1.9
+docker-workflow:1.13
+durable-task:1.16
+email-ext:2.61
+favorite:2.3.1
 form-element-path:1.8
-icon-shim:2.0.3
+git-client:2.6.0
+git-server:1.7
+git:3.6.4
+github-api:1.90
+github-branch-source:2.2.6
+github:1.28.1
+gradle:1.26
+handlebars:1.1.1
+hipchat:2.1.1
+htmlpublisher:1.14
+http_request:1.8.22
+infradna-backup:3.35
+jackson2-api:2.7.3
+jira:2.4.2
+jquery-detached:1.2.1
+jsch:0.1.54.1
 junit:1.21
+ldap:1.18
 mailer:1.20
 mapdb-api:1.0.9.0
-variant:1.1
-nectar-rbac:5.17
+matrix-auth:2.1.1
+matrix-project:1.11
+mercurial:2.2
+metrics:3.1.2.10
+momentjs:1.1.1
+nectar-license:8.7
+nectar-rbac:5.19
+node-iterator-api:1.5.0
 one-shot-executor:1.2
-operations-center-agent:2.73.0.1
-operations-center-analytics-config:2.73.0.2
-token-macro:2.1
-script-security:1.34
-unique-id:2.1.3
-operations-center-context:2.73.0.3
-operations-center-analytics-reporter:2.73.0.2
-operations-center-client:2.73.0.1
-tiger-tenant:1.0.14
-palace-cloud:2.0.2
-pse-cje-client-license:1.6.3
-aws-credentials:1.21
-workflow-step-api:2.13
-infradna-backup:3.34
-plain-credentials:1.4
-cloudbees-support:3.11
-credentials-binding:1.12
-cloudbees-monitoring:2.7
+operations-center-agent:2.89.0.1
+operations-center-analytics-config:2.89.0.1
+operations-center-analytics-reporter:2.89.0.1
+operations-center-client:2.89.0.1
+operations-center-cloud:2.89.0.1
+operations-center-context:2.89.0.1
+palace-cloud:2.0.5
+pipeline-build-step:2.5.1
+pipeline-graph-analysis:1.5
+pipeline-input-step:2.8
+pipeline-milestone-step:1.3.1
 pipeline-model-api:1.2.2
-blueocean-rest:1.3.0
-ace-editor:1.1
-cloudbees-template:4.33
-jquery-detached:1.2.1
+pipeline-model-declarative-agent:1.1.1
+pipeline-model-definition:1.2.2
+pipeline-model-extensions:1.2.2
+pipeline-rest-api:2.9
+pipeline-stage-step:2.2
+pipeline-stage-tags-metadata:1.2.2
 pipeline-stage-view:2.9
-workflow-scm-step:2.6
+pipeline-utility-steps:1.5.1
+plain-credentials:1.4
+pse-cje-client-license:1.6.3
 pubsub-light:1.12
 scm-api:2.2.2
-htmlpublisher:1.14
-workflow-api:2.22
-cloudbees-workflow-ui:2.1
-workflow-support:2.15
-blueocean-web:1.3.0
-workflow-cps:2.41
-blueocean-jwt:1.3.0
-workflow-job:2.14.1
-gradle:1.26
-blueocean:1.3.0
-pipeline-model-extensions:1.2.2
-pipeline-milestone-step:1.3.1
-ssh-credentials:1.13
-favorite:2.3.1
-git-client:2.5.0
-blueocean-config:1.3.0
-git-server:1.7
-node-iterator-api:1.5.0
-workflow-cps-global-lib:2.9
-mercurial:2.0
-branch-api:2.0.11
-pipeline-build-step:2.5.1
-workflow-multibranch:2.16
-workflow-aggregator:2.5
-authentication-tokens:1.3
-wikitext:3.8
-docker-commons:1.8
-operations-center-cloud:2.73.0.1
-workflow-durable-task-step:2.15
-ldap:1.17
-docker-workflow:1.13
-pipeline-graph-analysis:1.5
-pipeline-stage-tags-metadata:1.2.2
-cloudbees-ssh-slaves:2.1
-pipeline-model-declarative-agent:1.1.1
-git:3.6.0
-pipeline-stage-step:2.2
-github-api:1.86
-workflow-basic-steps:2.6
-aws-java-sdk:1.11.119
-pipeline-model-definition:1.2.2
-github:1.27.0
-cloudbees-nodes-plus:1.15
-cloudbees-groovy-view:1.6
-pipeline-input-step:2.8
-cloudbees-view-creation-filter:1.4
-pipeline-rest-api:2.9
-cloudbees-jsync-archiver:5.6
-github-branch-source:2.2.3
-matrix-project:1.11
-blueocean-jira:1.3.0
-email-ext:2.57.2
-cloudbees-bitbucket-branch-source:2.2.3
-cloudbees-workflow-template:2.7
-momentjs:1.1.1
-hipchat:2.1.1
-workflow-cps-checkpoint:2.4
-jira:2.4.2
-handlebars:1.1.1
-gitea:1.0.3
-pipeline-github-lib:1.0
-github-organization-folder:1.6
-blueocean-pipeline-scm-api:1.3.0
-blueocean-rest-impl:1.3.0
-blueocean-pipeline-api-impl:1.3.0
-blueocean-github-pipeline:1.3.0
-blueocean-git-pipeline:1.3.0
-blueocean-bitbucket-pipeline:1.3.0
-blueocean-personalization:1.3.0
-blueocean-display-url:2.1.0
+script-security:1.39
 sse-gateway:1.15
-blueocean-events:1.3.0
-blueocean-pipeline-editor:1.3.0
-blueocean-i18n:1.3.0
-blueocean-autofavorite:1.0.0
-blueocean-dashboard:1.3.0
-http_request:1.8.21
-slack:2.3
-pipeline-utility-steps:1.5.1
+ssh-credentials:1.13
+structs:1.10
+support-core:2.43
+tiger-tenant:1.0.15
 timestamper:1.8.8
+token-macro:2.1
+unique-id:2.1.3
+variant:1.1
+wikitext:3.8
+windows-slaves:1.3.1
+workflow-aggregator:2.5
+workflow-api:2.22
+workflow-basic-steps:2.6
+workflow-cps-checkpoint:2.4
+workflow-cps-global-lib:2.9
+workflow-cps:2.41
+workflow-durable-task-step:2.17
+workflow-job:2.15
+workflow-multibranch:2.16
+workflow-scm-step:2.6
+workflow-step-api:2.13
+workflow-support:2.16
